requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
requests==2.22.0
